refactor whole file program to be modular
The program is getting bigger that's why it need a better approach
to creating it, and the approach is to make it more modular through
diffrent class files.
